+++ package.json
@@ -17,6 +17,7 @@
     "chart.js": "^4.4.0",
     "esbuild": "^0.19.3",
     "formik": "^2.4.4",
+    "lodash.isequal": "^4.5.0",
     "modern-normalize": "^2.0.0",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
@@ -42,4 +43,3 @@
     "vite": "^4.4.5"
   }
 }
-
