+++ src/pages/DashboardPage/Chart/Chart.module.scss
@@ -1,0 +1,40 @@
+@import '@/scss/base/variables';
+@import '@/scss/base/mixins';
+
+.graphTooltip {
+  position: absolute;
+  width: 104px;
+  padding: 6px;
+  background-color: $primary-black-2;
+  box-shadow: 0px 4px 14px 0px rgba(227, 255, 168, 0.20);
+}
+
+.closeBtn {
+  padding: 0px;
+  display: flex;
+  align-items: center;
+  margin-left: 76px;
+  width: 16px;
+  height: 16px;
+  background-color: transparent;
+}
+
+.tooltipValue {
+  font-size: 32px;
+  font-weight: 500;
+  line-height: 1.19;
+}
+
+.tooltipTitle {
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 20px;
+  color: $primary-grey;
+  width: 100%;
+  text-align: center;
+}
+
+.tooltipField {
+  display: inline-flex;
+  flex-direction: column;
+}
