+++ src/pages/DashboardPage/Modal/Modal.jsx
@@ -1,4 +1,4 @@
-import styles from '../DashboardPage.module.scss';
+import styles from './Modal.module.scss'
 import PropTypes from 'prop-types';
 import { useEffect } from "react";
 
@@ -18,14 +18,8 @@ const Modal = ({onClose, children}) => {
     
   }, [onClose]);
 
-  const handleBackdropClick = e => {
-    if (e.currentTarget === e.target) {
-      onClose();
-    }
-  };
-
   return (
-    <div onClick={(e) => handleBackdropClick(e)} className={styles.modal}>
+    <div className={styles.modal}>
       {children}
     </div>
   );
